addressed all comments on PR (either by updating code, or via responses to those comments in-line) also moved ConstantStepWiseRateIncreaser.java  from es-plugin to core as per Ioannis's request from a while back.
