+++ ndbench-es-plugins/src/test/java/com/netflix/ndbench/plugin/es/AbstractPluginTest.java
@@ -2,7 +2,6 @@ package com.netflix.ndbench.plugin.es;
 
 import com.google.common.collect.ImmutableList;
 import com.netflix.ndbench.api.plugin.DataGenerator;
-import com.netflix.ndbench.api.plugin.builtin.IEsConfig;
 import com.netflix.ndbench.core.config.IConfiguration;
 import com.netflix.ndbench.core.discovery.IClusterDiscovery;
 import org.slf4j.Logger;
