+++ ndbench-es-plugins/src/test/java/com/netflix/ndbench/plugin/es/AutotuneTest.java
@@ -2,7 +2,7 @@ package com.netflix.ndbench.plugin.es;
 
 import com.google.common.collect.ImmutableList;
 import com.netflix.ndbench.api.plugin.NdBenchMonitor;
-import com.netflix.ndbench.api.plugin.builtin.IEsConfig;
+import com.netflix.ndbench.core.util.ConstantStepWiseRateIncreaser;
 import org.junit.Test;
 
 import java.util.ArrayList;
@@ -13,20 +13,7 @@ import static org.hamcrest.CoreMatchers.is;
 import static org.junit.Assert.assertThat;
 
 
-public class AutotuneTest extends AbstractPluginTest
-
-{
-    @Test
-    public void testGetRateReturnsProperRateRelativeToTimeZero() throws Exception {
-        ConstantStepWiseRateIncreaser increaser = new ConstantStepWiseRateIncreaser(100, 10, 0, 10);
-        assertThat(increaser.getRateForGivenClockTime(0, 9), is(equalTo(0.0)));
-        assertThat(increaser.getRateForGivenClockTime(0, 10), is(equalTo(1.0)));
-        assertThat(increaser.getRateForGivenClockTime(0, 99), is(equalTo(9.0)));
-        assertThat(increaser.getRateForGivenClockTime(0, 100), is(equalTo(10.0)));
-        assertThat(increaser.getRateForGivenClockTime(0, 101), is(equalTo(10.0)));
-        assertThat(increaser.getRateForGivenClockTime(0, Long.MAX_VALUE - 1), is(equalTo(10.0)));
-    }
-
+public class AutotuneTest extends AbstractPluginTest {
     @Test
     public void testRateStopsIncreasingAfterAcceptableWriteFailureThresholdReached() throws Exception {
